Update test suite for simplified data and model paths
