+++ tests/models/test_utils.py
@@ -10,6 +10,7 @@ from cell_observatory_platform.models.layers.utils import (
     point_sample,
     point_sample_labelmap,
     point_sample_labelmap_batched,
+    sample_box_points_from_boxes,
 )
 
 CUDA_AVAILABLE = torch.cuda.is_available()
@@ -252,3 +253,130 @@ def test_compute_unmasked_ratio_all_valid_and_partial():
 
     # sample 1 unchanged (all valid)
     assert torch.allclose(ratios_partial[1], torch.tensor([1.0, 1.0, 1.0], device=device))
+
+
+# ----------------------------------------------------------------------------- #
+# sample_box_points_from_boxes
+# ----------------------------------------------------------------------------- #
+
+
+def test_sample_box_points_from_boxes_xyzxyz_noise_zero_matches_corners():
+    # With noise=0, output points must equal the box corners exactly (clamped
+    # to image bounds). Labels must be [top_left_label, bottom_right_label].
+    device = torch.device("cpu")
+    Z, Y, X = 10, 20, 30
+    boxes = torch.tensor(
+        [[1.0, 2.0, 3.0, 5.0, 6.0, 7.0],
+         [10.0, 11.0, 5.0, 12.0, 13.0, 8.0]],
+        device=device,
+    )
+    pts, lbls = sample_box_points_from_boxes(
+        boxes=boxes, box_format="xyzxyz", image_shape=(Z, Y, X), noise=0.0,
+    )
+    assert pts.shape == (2, 2, 3)
+    assert lbls.shape == (2, 2)
+    # Row 0: top-left (1, 2, 3), bottom-right (5, 6, 7)
+    assert torch.equal(
+        pts[0], torch.tensor([[1.0, 2.0, 3.0], [5.0, 6.0, 7.0]])
+    )
+    assert torch.equal(lbls[0], torch.tensor([2, 3], dtype=torch.int32))
+
+
+def test_sample_box_points_from_boxes_zyxzyx_format_conversion():
+    # zyxzyx input must map to identical xyz output as the xyzxyz input case.
+    device = torch.device("cpu")
+    Z, Y, X = 10, 20, 30
+    boxes_zyx = torch.tensor(
+        [[3.0, 2.0, 1.0, 7.0, 6.0, 5.0]],  # (z1, y1, x1, z2, y2, x2)
+        device=device,
+    )
+    boxes_xyz = torch.tensor(
+        [[1.0, 2.0, 3.0, 5.0, 6.0, 7.0]],  # (x1, y1, z1, x2, y2, z2)
+        device=device,
+    )
+    pts_zyx, _ = sample_box_points_from_boxes(
+        boxes=boxes_zyx, box_format="zyxzyx", image_shape=(Z, Y, X), noise=0.0,
+    )
+    pts_xyz, _ = sample_box_points_from_boxes(
+        boxes=boxes_xyz, box_format="xyzxyz", image_shape=(Z, Y, X), noise=0.0,
+    )
+    assert torch.equal(pts_zyx, pts_xyz)
+
+
+def test_sample_box_points_from_boxes_cxcyczwhd_format_conversion():
+    # cxcyczwhd input must map to identical xyz output as the xyzxyz input case.
+    device = torch.device("cpu")
+    Z, Y, X = 10, 20, 30
+    # Box centered at (3, 4, 5) (xyz) with size (2, 4, 6) (whd)
+    # -> x1=2, y1=2, z1=2, x2=4, y2=6, z2=8
+    boxes_c = torch.tensor([[3.0, 4.0, 5.0, 2.0, 4.0, 6.0]], device=device)
+    boxes_x = torch.tensor([[2.0, 2.0, 2.0, 4.0, 6.0, 8.0]], device=device)
+    pts_c, _ = sample_box_points_from_boxes(
+        boxes=boxes_c, box_format="cxcyczwhd", image_shape=(Z, Y, X), noise=0.0,
+    )
+    pts_x, _ = sample_box_points_from_boxes(
+        boxes=boxes_x, box_format="xyzxyz", image_shape=(Z, Y, X), noise=0.0,
+    )
+    assert torch.equal(pts_c, pts_x)
+
+
+def test_sample_box_points_from_boxes_clamps_to_image_bounds():
+    # Boxes outside the image must be clamped to [0, dim-1].
+    device = torch.device("cpu")
+    Z, Y, X = 10, 20, 30
+    boxes = torch.tensor(
+        [[-5.0, -5.0, -5.0, 100.0, 100.0, 100.0]],
+        device=device,
+    )
+    pts, _ = sample_box_points_from_boxes(
+        boxes=boxes, box_format="xyzxyz", image_shape=(Z, Y, X), noise=0.0,
+    )
+    assert torch.equal(pts[0, 0], torch.tensor([0.0, 0.0, 0.0]))
+    assert torch.equal(pts[0, 1], torch.tensor([X - 1.0, Y - 1.0, Z - 1.0]))
+
+
+def test_sample_box_points_from_boxes_valid_zeros_pad_rows():
+    # Pad rows (valid=False) must get all-zero points and label 0.
+    device = torch.device("cpu")
+    Z, Y, X = 10, 20, 30
+    boxes = torch.tensor(
+        [[1.0, 2.0, 3.0, 5.0, 6.0, 7.0],
+         [10.0, 11.0, 5.0, 12.0, 13.0, 8.0]],
+        device=device,
+    )
+    valid = torch.tensor([True, False])
+    pts, lbls = sample_box_points_from_boxes(
+        boxes=boxes, box_format="xyzxyz", image_shape=(Z, Y, X), noise=0.0,
+        valid=valid,
+    )
+    assert torch.equal(lbls[0], torch.tensor([2, 3], dtype=torch.int32))
+    assert torch.equal(lbls[1], torch.tensor([0, 0], dtype=torch.int32))
+    assert torch.all(pts[1] == 0.0)
+    assert not torch.all(pts[0] == 0.0)
+
+
+def test_sample_box_points_from_boxes_noise_bounded():
+    # With noise>0, perturbed corners must still lie inside the image and
+    # within noise_bound pixels of the original (per-axis).
+    torch.manual_seed(0)
+    device = torch.device("cpu")
+    Z, Y, X = 50, 100, 200
+    boxes = torch.tensor(
+        [[10.0, 20.0, 30.0, 40.0, 50.0, 45.0]] * 10,
+        device=device,
+    )
+    pts, _ = sample_box_points_from_boxes(
+        boxes=boxes, box_format="xyzxyz", image_shape=(Z, Y, X),
+        noise=0.1, noise_bound=20,
+    )
+    assert torch.all(pts[..., 0] >= 0) and torch.all(pts[..., 0] <= X - 1)
+    assert torch.all(pts[..., 1] >= 0) and torch.all(pts[..., 1] <= Y - 1)
+    assert torch.all(pts[..., 2] >= 0) and torch.all(pts[..., 2] <= Z - 1)
+
+
+def test_sample_box_points_from_boxes_rejects_bad_shape():
+    boxes = torch.zeros((3, 4))  # last dim should be 6
+    with pytest.raises(ValueError):
+        sample_box_points_from_boxes(
+            boxes=boxes, box_format="xyzxyz", image_shape=(4, 5, 6),
+        )
